dsagent-main-wopure/requirements-dev.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiolimiter==1.1.0
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.18.1
anyio==4.4.0
anytree==2.12.1
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
asyncer==0.0.7
attrs==23.2.0
azure-core==1.30.2
azure-identity==1.17.1
backoff==2.2.1
bce-python-sdk==0.9.17
bcrypt==4.1.3
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
bs4==0.0.2
build==1.2.1
cachetools==5.3.3
camel-converter==3.1.2
certifi==2024.6.2
cffi==1.16.0
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.3.2
chevron==0.14.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cryptography==42.0.8
cycler==0.12.1
dashscope==1.14.1
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
Django==5.0.6
dnspython==2.6.1
elastic-transport==8.13.1
elasticsearch==8.14.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.115.5
fastjsonschema==2.20.0
filelock==3.15.4
filetype==1.2.0
fire==0.4.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gitdb==4.0.11
gitignore_parser==0.1.9
GitPython==3.1.40
google-ai-generativelanguage==0.4.0
google-api-core==2.19.1
google-auth==2.30.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
grpcio-tools==1.62.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
imap-tools==1.5.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.27.1
ipython==8.17.2
ipywidgets==8.1.1
isodate==0.6.1
jedi==0.19.1
jieba==0.42.1
joblib==1.4.2
jsonschema==4.22.0
jsonschema-path==0.3.3
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kaggle==1.6.17
kiwisolver==1.4.7
kubernetes==30.1.0
lancedb==0.4.0
Lazify==0.4.0
lazy-object-proxy==1.10.0
libcst==1.0.1
literalai==0.0.623
llama-cloud==0.0.9
llama-index==0.10.15
llama-index-agent-openai==0.1.7
llama-index-cli==0.1.12
llama-index-core==0.10.15
llama-index-embeddings-azure-openai==0.1.6
llama-index-embeddings-openai==0.1.5
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-azure-openai==0.1.4
llama-index-llms-openai==0.1.8
llama-index-multi-modal-llms-openai==0.1.7
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.4
llama-index-readers-llama-parse==0.1.6
llama-index-retrievers-bm25==0.1.3
llama-index-vector-stores-chroma==0.1.6
llama-index-vector-stores-elasticsearch==0.1.6
llama-index-vector-stores-faiss==0.1.1
llama-parse==0.5.14
llamaindex-py-client==0.1.19
loguru==0.6.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
meilisearch==0.21.0
mmh3==4.1.0
monotonic==1.6
more-itertools==10.3.0
motor==3.5.0
mpmath==1.3.0
msal==1.29.0
msal-extensions==1.2.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.9.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.24.3
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.6.1
openapi-core==0.18.2
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.6
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.1.1
parse==1.20.2
parso==0.8.4
pathable==0.4.3
patsy==1.0.1
pillow==10.3.0
platformdirs==4.2.2
playwright==1.44.0
portalocker==2.10.0
posthog==3.5.0
prance==23.6.21.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.5
psutil==6.0.0
pulsar-client==3.5.0
pure-eval==0.2.2
py==1.11.0
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pyee==11.1.0
Pygments==2.18.0
PyJWT==2.8.0
pylance==0.9.0
pymongo==4.8.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyparsing==3.2.0
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.0
python-engineio==4.10.1
python-multipart==0.0.9
python-slugify==8.0.4
python-socketio==5.11.4
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qdrant-client==1.7.0
qianfan==0.3.2
rank-bm25==0.2.2
ratelimiter==1.2.0.post0
referencing==0.30.2
regex==2023.12.25
requests==2.32.3
requests-oauthlib==2.0.0
retry==0.9.2
rfc3339-validator==0.1.4
rich==13.6.0
rpds-py==0.18.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.3.2
scipy==1.14.0
seaborn==0.13.2
selenium==4.25.0
semantic-kernel==0.4.3.dev0
semver==3.0.2
simple-websocket==1.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.41.2
statsmodels==0.14.4
sympy==1.13.0
syncer==2.0.3
ta==0.10.2
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.6.0
tokenizers==0.19.1
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typer==0.9.0
typing-inspect==0.8.0
typing_extensions==4.9.0
tzdata==2024.1
uptrace==1.27.0
urllib3==1.26.19
uvicorn==0.25.0
watchfiles==0.20.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.6.2
websockets==11.0.3
Werkzeug==3.0.3
widgetsnbextension==4.0.11
win32-setctime==1.1.0
wrapt==1.15.0
wsproto==1.2.0
xgboost==2.1.0
yarl==1.9.4
zhipuai==2.0.1
zipp==3.19.2
